Restructured simulation170925 to AlfvenLUT - a derived class of Simulation that incorporates LUT stuff.  Also prepared the way for calculated Alfven waves, but more work is needed on the code.  Right now, a CUDA compile option: device relocatable code is needed to compile correctly.  I don't like this and will attempt to change later.  Worst case, I'll work it into a CMakeLists if possible...

+++ SimTools/include/ParticleClass/Particle.h
@@ -43,7 +43,7 @@ public:
 		if (attributeNames.size() < numberOfVelocityDims_m + numberOfPositionDims_m)
 		{
 			std::cout << "Warning: Particle: not enough attribute names specified.  Generic names being generated.  If you try to load bin files into the data array, it probably won't work.\n";
-			for (int diff = 0; diff = numberOfVelocityDims_m + numberOfPositionDims_m - attributeNames.size(); diff++)
+			for (int diff = 0; diff = numberOfVelocityDims_m + numberOfPositionDims_m - static_cast<int>(attributeNames.size()); diff++)
 				attributeNames_m.push_back(std::to_string(diff) + ".bin");
 		}
 	}
